requirements.txt
flask==2.0.3
werkzeug==2.0.3
python-dotenv==0.19.0
requests==2.26.0
google-api-python-client==2.48.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6